requirements.txt
smolagents>=0.1.0
gradio>=4.0.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
tabulate>=0.9.0 
smolagents[transformers]
langchain_community
langchain-openai>=0.0.2
pytesseract
langchain_experimental
wikipedia
arxiv
openai>=1.0.0
openpyxl 
tesseract
pymupdf 
timm